Moved {Tone, AbsoluteTone} to tone.
